fixed urls detail contracts
