+++ document_management/apps/contracts/urls.py
@@ -1,7 +1,13 @@
-from django.urls import path, include
+from django.urls import path
+from . import views
 
 app_name = 'contracts'
 
 urlpatterns = [
-    path('constructions/', include('document_management.apps.contracts.constructions.urls', namespace="constructions")),
+    path('', views.index, name="index"),
+    path('add/', views.add, name="add"),
+    path('<int:id>/edit/', views.edit, name="edit"),
+    path('<int:id>/delete/', views.delete, name="delete"),
+    path('<int:id>/details/', views.details, name="details"),
+    path('<int:id>/upload/', views.upload, name="upload"),
 ]
