+++ templates/addendums/details.html
@@ -91,7 +91,7 @@
                         <div class="col-md-6">
                             <div class="form-group mb-3">
                                 <h6 class="font-14">Nomor MoU / Contract :</h6>
-                                <p><a href="{% url 'backoffice:contracts:constructions:details' 1 %}"># K00-0001</a></p>
+                                <p><a href="{% url 'backoffice:contracts:details' 1 %}"># K00-0001</a></p>
                             </div>
                         </div>
                     </div>
