+++ templates/contracts/delete.html
@@ -12,8 +12,7 @@
                     <ol class="breadcrumb m-0">
                         <li class="breadcrumb-item"><a href="{% url 'backoffice:dashboard' %}">DMS</a></li>
                         <li class="breadcrumb-item"><a href="javascript: void(0);">Contracts</a></li>
-                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:constructions:index' %}">Constructions</a></li>
-                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:constructions:details' 1 %}">Detail</a></li>
+                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:details' 1 %}">Detail</a></li>
                         <li class="breadcrumb-item active">Delete</li>
                     </ol>
                 </div>
