+++ templates/contracts/edit.html
@@ -12,12 +12,11 @@
                     <ol class="breadcrumb m-0">
                         <li class="breadcrumb-item"><a href="{% url 'backoffice:dashboard' %}">DMS</a></li>
                         <li class="breadcrumb-item"><a href="javascript: void(0);">Contracts</a></li>
-                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:constructions:index' %}">Constructions</a></li>
-                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:constructions:details' 1 %}">Detail</a></li>
+                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:details' 1 %}">Detail</a></li>
                         <li class="breadcrumb-item active">Edit</li>
                     </ol>
                 </div>
-                <h4 class="page-title">Edit Construction</h4>
+                <h4 class="page-title">Edit Contract</h4>
             </div>
         </div>
     </div>
