+++ templates/contracts/upload.html
@@ -12,12 +12,11 @@
                     <ol class="breadcrumb m-0">
                         <li class="breadcrumb-item"><a href="{% url 'backoffice:dashboard' %}">DMS</a></li>
                         <li class="breadcrumb-item"><a href="javascript: void(0);">Contracts</a></li>
-                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:constructions:index' %}">Constructions</a></li>
-                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:constructions:details' 1 %}">Details</a></li>
+                        <li class="breadcrumb-item"><a href="{% url 'backoffice:contracts:details' 1 %}">Details</a></li>
                         <li class="breadcrumb-item active">Upload</li>
                     </ol>
                 </div>
-                <h4 class="page-title">Construction Upload</h4>
+                <h4 class="page-title">Upload Contract</h4>
             </div>
         </div>
     </div>
